load character from model character
